Add Core.Test MSTest project and CandidateHelperTests for CandidateHelper class

+++ Core/Entities/CandidateSource.cs
@@ -5,7 +5,7 @@ namespace Core.Entities
   public class CandidateSource
   {
     [JsonProperty("candidateId")]
-    public long CandidateId { get; set; }
+    public int CandidateId { get; set; }
 
     [JsonProperty("name")]
     public string Name { get; set; }
